Move Quiz* models to own app
