+++ backend/api/views.py
@@ -12,13 +12,12 @@ from rest_framework.decorators import api_view
 from rest_framework.response import Response
 
 from api import constants, utilities_sendinblue
-from api.filters import QuestionFilter, QuizFilter
-from api.models import Question, Quiz
+from api.filters import QuestionFilter
+from api.models import Question
 from api.serializers import (
     QuestionDifficultyChoiceSerializer,
     QuestionFullStringSerializer,
     QuestionSerializer,
-    QuizSerializer,
     SimpleChoiceSerializer,
 )
 
@@ -146,20 +145,6 @@ def author_list(request):
     )
 
     return Response(list(question_authors))
-
-
-class QuizViewSet(mixins.ListModelMixin, mixins.RetrieveModelMixin, viewsets.GenericViewSet):
-    queryset = Quiz.objects.published()
-    serializer_class = QuizSerializer
-    filter_class = QuizFilter
-
-    @extend_schema(summary="Lister tous les quiz *publiés*", tags=[Quiz._meta.verbose_name_plural])
-    def list(self, request, *args, **kwargs):
-        return super().list(request, args, kwargs)
-
-    @extend_schema(summary="Détail d'un quiz *publié*", tags=[Quiz._meta.verbose_name_plural])
-    def retrieve(self, request, *args, **kwargs):
-        return super().retrieve(request, args, kwargs)
 
 
 @api_view(["GET", "POST"])
