+++ backend/stats/models.py
@@ -4,7 +4,8 @@ from django.db import models
 from django.db.models import Count, Sum
 
 from api import constants as api_constants
-from api.models import Question, Quiz
+from api.models import Question
+from quizs.models import Quiz
 from stats import constants
